+++ src/main.rs
@@ -2,20 +2,22 @@
 //!
 //! [![badge github]][url github]
 //! [![badge crates.io]][url crates.io]
-//! [![badge docs.rs]][url docs.rs]
 //! [![badge license]][url license]
 //!
 //! [badge github]: https://img.shields.io/badge/github-FloGa%2Fnormalized--hasher-green
 //! [badge crates.io]: https://img.shields.io/crates/v/normalized-hasher
-//! [badge docs.rs]: https://img.shields.io/docsrs/normalized-hasher
 //! [badge license]: https://img.shields.io/crates/l/normalized-hasher
 //!
 //! [url github]: https://github.com/FloGa/normalized-hasher
 //! [url crates.io]: https://crates.io/crates/normalized-hasher
-//! [url docs.rs]: https://docs.rs/normalized-hasher
 //! [url license]: https://github.com/FloGa/normalized-hasher/blob/develop/LICENSE
 //!
 //! Create cross-platform hashes of text files.
+//!
+//! *This is the binary crate. If you're looking for the library crate instead, go
+//! to [`normalized-hash`].*
+//!
+//! [`normalized-hash`]: https://github.com/FloGa/normalized-hasher/crates/normalized-hash
 //!
 //! ## Motivation
 //!
@@ -48,7 +50,7 @@
 //!
 //! `normalized-hasher` can be installed easily through Cargo via `crates.io`:
 //!
-//! ```shell script
+//! ```shell
 //! cargo install --locked normalized-hasher
 //! ```
 //!
@@ -67,24 +69,86 @@
 //! ## Usage
 //!
 //! ```text
-//! Usage: normalized-hasher <FILE_IN> [FILE_OUT]
+//! Usage: normalized-hasher [OPTIONS] <FILE_IN> [FILE_OUT]
 //!
 //! Arguments:
-//!   <FILE_IN>   File to be hashed
-//!   [FILE_OUT]  Optional file path to write normalized input into
+//!   <FILE_IN>
+//!           File to be hashed
+//!
+//!   [FILE_OUT]
+//!           Optional file path to write normalized input into
 //!
 //! Options:
-//!   -h, --help     Print help
-//!   -V, --version  Print version
+//!       --eol <EOL>
+//!           End-of-line sequence, will be appended to each normalized line for hashing
+//!
+//!           [default: "\n"]
+//!
+//!       --ignore-whitespaces
+//!           Ignore all whitespaces
+//!
+//!           This will remove all whitespaces from the input file when generating the hash.
+//!
+//!       --no-eof
+//!           Skip last end-of-line on end-of-file
+//!
+//!           With this flag, no trailing EOL will be appended at the end of the file.
+//!
+//!   -h, --help
+//!           Print help (see a summary with '-h')
+//!
+//!   -V, --version
+//!           Print version
+//! ```
+//!
+//! ### Flags
+//!
+//! -   `--eol`
+//!
+//!     With the `--eol` flag you can change the end-of-line sequence that will be
+//!     appended to each normalized line to generate the hash. This can be useful
+//!     if you explicitly want CRLF endings, for example.
+//!
+//!     Please note that you need to escape control characters properly in your
+//!     shell. For Bash, you can type:
+//!
+//!     ```shell
+//!     normalized-hasher --eol $'\r\n' input.txt output.txt
+//!     ```
+//!
+//! -   `--ignore-whitespaces`
+//!
+//!     In some extreme cases, you might want to ignore all whitespaces in a file.
+//!     With the `--ignore-whitespaces` flag, all whitespaces are removed prior to
+//!     generate the hash.
+//!
+//! -   `--no-eof`
+//!
+//!     With the `--no-eof` flag you can avoid appending the EOL sequence at the
+//!     end of the file. This is for use cases where such trailing EOL is not
+//!     desireable, like in Windows files. In contrast to UNIX files which usually
+//!     end with a final LF, Windows files do not usually end with an additional
+//!     CRLF.
+//!
+//! ## Examples
+//!
+//! Simple example with default options, without writing an output file:
+//!
+//! ```shell
+//! normalized-hasher input.txt
+//! ```
+//!
+//! More complex example, with writing output:
+//!
+//! ```shell
+//! normalized-hasher --eol $'\r\n' --no-eof input.txt output.txt
 //! ```
 
 use std::ffi::OsString;
-use std::fs::File;
-use std::io::{BufRead, BufReader, BufWriter, Write};
-use std::path::Path;
 
 use clap::Parser;
-use sha2::{Digest, Sha256};
+
+use normalized_hash::Hasher;
 
 #[derive(Parser)]
 #[clap(author, version, about, long_about = None)]
@@ -94,91 +158,43 @@ struct Cli {
 
     /// Optional file path to write normalized input into
     file_out: Option<OsString>,
-}
 
-fn hash_file(file_in: impl AsRef<Path>, file_out: Option<impl AsRef<Path>>) -> String {
-    let file_in = File::open(file_in).unwrap();
-    let file_in = BufReader::new(file_in);
+    /// End-of-line sequence, will be appended to each normalized line for hashing
+    #[arg(long, default_value = "\n")]
+    eol: String,
 
-    let mut file_out = match file_out {
-        Some(file_out) => {
-            let file_out = File::create(file_out).unwrap();
-            let file_out = BufWriter::new(file_out);
-            Some(file_out)
-        }
-        None => None,
-    };
+    /// Ignore all whitespaces
+    ///
+    /// This will remove all whitespaces from the input file when generating the hash.
+    #[arg(long)]
+    ignore_whitespaces: bool,
 
-    let mut hasher = Sha256::new();
-    for line in file_in.lines() {
-        let line = line.unwrap();
-        let line = format!("{}\n", line);
-        hasher.update(&line);
-
-        if let Some(file_out) = &mut file_out {
-            file_out.write_all(line.as_bytes()).unwrap();
-        }
-    }
-
-    let hash = hasher.finalize();
-
-    base16ct::lower::encode_string(&hash)
+    /// Skip last end-of-line on end-of-file
+    ///
+    /// With this flag, no trailing EOL will be appended at the end of the file.
+    #[arg(long)]
+    no_eof: bool,
 }
 
 fn main() {
     let cli = Cli::parse();
 
-    println!("{}", hash_file(cli.file_in, cli.file_out));
+    println!(
+        "{}",
+        Hasher::new()
+            .eol(cli.eol)
+            .no_eof(cli.no_eof)
+            .hash_file(cli.file_in, cli.file_out)
+    );
 }
 
 #[cfg(test)]
 mod tests {
-    use std::error::Error;
-    use std::fs;
-
-    use tempfile;
-
     use super::*;
 
     #[test]
     fn verify_app() {
         use clap::CommandFactory;
         Cli::command().debug_assert()
-    }
-
-    #[test]
-    fn check_empty_file() -> Result<(), Box<dyn Error>> {
-        let file = tempfile::NamedTempFile::new()?;
-
-        // Sanity check between hasher versions
-        let hash_expected = "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855";
-        let hash_actual = hash_file(file, None::<OsString>);
-
-        assert_eq!(hash_actual, hash_expected);
-
-        Ok(())
-    }
-
-    #[test]
-    fn check_different_eols() -> Result<(), Box<dyn Error>> {
-        let mut file_with_lf = tempfile::NamedTempFile::new()?;
-        let mut file_with_crlf = tempfile::NamedTempFile::new()?;
-
-        let file_with_lf_normalized = tempfile::NamedTempFile::new()?;
-        let file_with_crlf_normalized = tempfile::NamedTempFile::new()?;
-
-        file_with_lf.write_all("A\nb".as_ref())?;
-        file_with_crlf.write_all("A\r\nb".as_ref())?;
-
-        let hash_with_lf = hash_file(file_with_lf, Some(&file_with_lf_normalized));
-        let hash_with_crlf = hash_file(file_with_crlf, Some(&file_with_crlf_normalized));
-
-        assert_eq!(hash_with_lf, hash_with_crlf);
-        assert_eq!(
-            fs::read_to_string(file_with_lf_normalized)?,
-            fs::read_to_string(file_with_crlf_normalized)?
-        );
-
-        Ok(())
     }
 }